# services/engine/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chatbot Engine API
  description: |
    REST API for the Chatbot Engine — an NLP-powered chatbot platform that supports
    multi-group configuration, intent recognition, query execution, learning feedback,
    and administrative management.
  version: 1.0.0
  contact:
    name: Chatbot Engine Team
servers:
  - url: http://localhost:4000
    description: Local development
  - url: '{productionUrl}'
    description: Production
    variables:
      productionUrl:
        default: https://chatbot-engine.example.com
        description: Production server URL
...
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API key required for all `/api/admin/*` routes when the `ENGINE_API_KEY`
        environment variable is set. Can also be passed as a Bearer token in the
        Authorization header.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Internal server error
...
    QuerySummary:
      type: object
      properties:
        name:
          type: string
          example: open_tickets
        description:
          type: string
          example: Returns all open support tickets
        filters:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [api, url, document, csv]
          example: api
...
    AdminGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        sources:
          type: array
          items:
            type: string
        apiBaseUrl:
          type: string
          nullable: true
        hasCorpus:
          type: boolean
        hasFaq:
          type: boolean
        hasTemplates:
          type: boolean
...
    IntentItem:
      type: object
      properties:
        intent:
          type: string
          example: greeting
        utterances:
          type: array
          items:
            type: string
          example: [hello, hi there, hey]
        answers:
          type: array
          items:
            type: string
        utteranceCount:
          type: integer
...
paths:
...
  /api/queries:
    get:
      tags: [Queries]
      summary: List available queries for a group
      parameters:
        - name: groupId
          in: query
          schema:
            type: string
            default: default
          description: Group whose queries to list
      responses:
        '200':
          description: List of queries
          content:
            application/json:
              schema:
                type: object
                properties:
                  queries:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuerySummary'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  queries:
                    type: array
                    items: {}
...
  /api/admin/groups:
    get:
      tags: [Admin/Groups]
      summary: List all groups with configuration details
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminGroup'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items: {}
...
  /api/admin/queries:
    get:
      tags: [Admin/Queries]
      summary: List all query definitions
      security:
        - ApiKeyAuth: []
      parameters:
        - name: source
          in: query
          schema:
            type: string
          description: Comma-separated list of sources to filter by
      responses:
        '200':
          description: List of queries
          content:
            application/json:
              schema:
                type: object
                properties:
                  queries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminQuery'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  queries:
                    type: array
                    items: {}

    post:
      tags: [Admin/Queries]
      summary: Create a new query definition
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQueryRequest'
      responses:
        '201':
          description: Query created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminQuery'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Query name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Admin/Queries]
      summary: Update a query definition
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  example: q1
                name:
                  type: string
                description:
                  type: string
                source:
                  type: string
                url:
                  type: string
                type:
                  type: string
                filters:
                  type: array
                  items:
                    type: string
                estimatedDuration:
                  type: integer
                filePath:
                  type: string
                endpoint:
                  type: string
      responses:
        '200':
          description: Updated query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminQuery'
        '400':
          description: Missing id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Query not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Queries]
      summary: Delete a query definition
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Query ID to delete
      responses:
        '200':
          description: Query deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deletedQueryId:
                    type: string
        '400':
          description: Missing id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Query not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — FILTERS
  # ════════════════════════════════════════════════════════════════════════
  /api/admin/filters:
    get:
      tags: [Admin/Filters]
      summary: Get the filter configuration
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Filter config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterConfig'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Filters]
      summary: Create or update a filter
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFilterRequest'
      responses:
        '200':
          description: Filter saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  label:
                    type: string
                  type:
                    type: string
                  options:
                    type: array
                    items:
                      type: string
                  placeholder:
                    type: string
                    nullable: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Filters]
      summary: Delete a filter
      security:
        - ApiKeyAuth: []
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
          description: Filter key to delete
      responses:
        '200':
          description: Filter deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deletedKey:
                    type: string
        '400':
          description: Missing key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Filter not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — INTENTS
  # ════════════════════════════════════════════════════════════════════════
  /api/admin/intents:
    get:
      tags: [Admin/Intents]
      summary: List all intents with utterances and entities
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Intents and entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  intents:
                    type: array
                    items:
                      $ref: '#/components/schemas/IntentItem'
                  entities:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        options:
                          type: object
                          additionalProperties:
                            type: array
                            items:
                              type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Intents]
      summary: Create or update an intent
      description: If the intent already exists, its utterances and answers are replaced.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntentRequest'
      responses:
        '200':
          description: Intent saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  intent:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Intents]
      summary: Delete an intent
      security:
        - ApiKeyAuth: []
      parameters:
        - name: intent
          in: query
          required: true
          schema:
            type: string
          description: Intent name to delete
      responses:
        '200':
          description: Intent deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: string
        '400':
          description: Missing intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — ENTITIES
  # ════════════════════════════════════════════════════════════════════════
  /api/admin/entities:
    get:
      tags: [Admin/Entities]
      summary: List all entities
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Entities map
          content:
            application/json:
              schema:
                type: object
                properties:
                  entities:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        options:
                          type: object
                          additionalProperties:
                            type: array
                            items:
                              type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Entities]
      summary: Create or update an entity synonym set
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntityRequest'
      responses:
        '200':
          description: Entity saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  entityType:
                    type: string
                  optionKey:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Entities]
      summary: Delete an entity synonym set
      security:
        - ApiKeyAuth: []
      parameters:
        - name: entityType
          in: query
          required: true
          schema:
            type: string
        - name: optionKey
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entity deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: string
        '400':
          description: Missing parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — TEMPLATES
  # ════════════════════════════════════════════════════════════════════════
  /api/admin/templates:
    get:
      tags: [Admin/Templates]
      summary: Get all response templates (base and per-group)
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  baseTemplates:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
                  groupTemplates:
                    type: object
                    additionalProperties:
                      type: object
                      additionalProperties:
                        type: array
                        items:
                          type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Templates]
      summary: Create or update a response template
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '200':
          description: Template saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scope:
                    type: string
                  intent:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found (when scope is a group ID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Templates]
      summary: Delete a response template
      security:
        - ApiKeyAuth: []
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            default: base
          description: '"base" or a group ID'
        - name: intent
          in: query
          required: true
          schema:
            type: string
          description: Intent name to delete template for
      responses:
        '200':
          description: Template deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: string
        '400':
          description: Missing intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — USERS
  # ════════════════════════════════════════════════════════════════════════
  /api/admin/users:
    get:
      tags: [Admin/Users]
      summary: List all admin/viewer users
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: {}

    post:
      tags: [Admin/Users]
      summary: Create a new user
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/files:
    get:
      tags: [Admin/Files]
      summary: List knowledge-base files
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/KnowledgeFile'
                  totalFiles:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Files]
      summary: Upload a knowledge-base file
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileRequest'
      responses:
        '200':
          description: File created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  name:
                    type: string
        '400':
          description: Missing name or content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin/Files]
      summary: Delete a knowledge-base file
      security:
        - ApiKeyAuth: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Filename to delete
      responses:
        '200':
          description: File deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: string
        '400':
          description: Missing name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/settings:
    get:
      tags: [Admin/Settings]
      summary: Get current engine settings
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Engine settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    $ref: '#/components/schemas/EngineSettings'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin/Settings]
      summary: Update engine settings
      description: Merges the provided fields into the current settings.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineSettings'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  config:
                    $ref: '#/components/schemas/EngineSettings'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — LOGS
  # ════════════════════════════════════════════════════════════════════════
...
  /api/admin/audit:
    get:
      tags: [Admin/Audit]
      summary: Get admin audit log
      security:
        - ApiKeyAuth: []
      parameters:
        - name: resource
          in: query
          schema:
            type: string
          description: Filter by resource type (group, query, user, etc.)
        - name: groupId
          in: query
          schema:
            type: string
          description: Filter by group ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only include entries after this timestamp
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
                  count:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'