# services/engine/src/routes/admin/index.ts
import { Router, Request, Response } from 'express';
import groupsRouter from './groups';
import queriesRouter from './queries';
import filtersRouter from './filters';
import { intentsRouter, entitiesRouter } from './intents';
import templatesRouter from './templates';
import usersRouter from './users';
import filesRouter from './files';
import learningRouter from './learning';
import { settingsRouter, logsRouter } from './settings';
import { getAuditLog } from '@/lib/audit-logger';
import { Role, isValidRole } from '@/lib/rbac';
import { requirePermission } from '@/middleware/rbac';

const router = Router();

// Auth middleware — when ENGINE_API_KEY is set, require it on all admin routes
const engineApiKey = process.env.ENGINE_API_KEY;
if (engineApiKey) {
  router.use((req: Request, res: Response, next) => {
    const provided = req.headers['x-api-key'] || req.headers['authorization']?.replace('Bearer ', '');
    if (provided !== engineApiKey) {
      return res.status(401).json({ error: 'Unauthorized — valid API key required for admin access' });
    }
    next();
  });
}

// RBAC middleware — extract user role from X-User-Role header (set by the Next.js proxy)
// and attach it to req.userRole for downstream permission checks
router.use((req: Request, _res: Response, next) => {
  const roleHeader = req.headers['x-user-role'] as string | undefined;
  if (roleHeader && isValidRole(roleHeader)) {
    (req as any).userRole = roleHeader as Role;
  }
  next();
});

router.use('/groups', groupsRouter);
router.use('/queries', queriesRouter);
router.use('/filters', filtersRouter);
router.use('/intents', intentsRouter);
router.use('/entities', entitiesRouter);
router.use('/templates', templatesRouter);
router.use('/users', usersRouter);
router.use('/files', filesRouter);
router.use('/learning', learningRouter);
router.use('/settings', settingsRouter);
router.use('/logs', logsRouter);

// === AUDIT LOG ===
router.get('/audit', requirePermission('audit.view'), async (req: Request, res: Response) => {
  try {
    const resource = req.query.resource as string | undefined;
    const groupId = req.query.groupId as string | undefined;
    const limit = req.query.limit ? parseInt(req.query.limit as string, 10) : 50;
    const since = req.query.since as string | undefined;
    const entries = await getAuditLog({ resource, groupId, limit, since });
    return res.json({ entries, count: entries.length });
  } catch (err) {
    return res.status(500).json({ error: String(err) });
  }
});

export const adminRouter = router;
export default router;